refactor: Request DTO 내부 필드를 관심사에 따라 묶도록 변경
