+++ src/main/java/roomescape/payment/PaymentConfirmWebRequest.java
@@ -1,0 +1,12 @@
+package roomescape.payment;
+
+public record PaymentConfirmWebRequest(
+        String paymentKey,
+        String orderId,
+        int amount
+) {
+
+    public PaymentConfirmRequest toPaymentConfirmRequest() {
+        return new PaymentConfirmRequest(paymentKey, orderId, amount);
+    }
+}
